requirements.txt
snowflake-connector-python>=2.7.0
python-dotenv>=0.19.0
mcp>=1.0.0
cryptography>=36.0.0